requirements.txt
pandas
numpy
matplotlib.pyplot
streamlit
pandasql